changed file names to match design
